requirements.txt
paho-mqtt==1.3.0
telegram==0.0.1
urllib3==1.22
python-telegram-bot
cherrypy
